refactor(cli): unify retry policy in one shared module
Extract the transient-network classifier and a createTransientRetryPolicy
factory into retryPolicy.ts so both transports share one definition instead of
each building the policy literal inline (and WebSocketLlmBackend importing the
classifier from ServerLlmBackend). Both backends now call
createTransientRetryPolicy(); index.tsx and the relocated classifier test import
from retryPolicy. This is the parity guard - the retry/retryable rules can no
longer drift between the two paths.

Refs #229
